refactor: adjust erda-cli config and project-deployment logic
