+++ tools/cli/cmd/config_delete.go
@@ -15,13 +15,10 @@
 package cmd
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
 
-	"github.com/erda-project/erda/apistructs"
-	"github.com/erda-project/erda/pkg/http/httputil"
 	"github.com/erda-project/erda/tools/cli/command"
+	"github.com/erda-project/erda/tools/cli/common"
 	"github.com/erda-project/erda/tools/cli/utils"
 )
 
@@ -30,63 +27,25 @@ var CONFIGDELETE = command.Command{
 	ParentName: "CONFIG",
 	ShortHelp:  "delete project workspace config",
 	Example: `
-  $ erda-cli config delete --orgName xxx --projectName yyy  --workspace DEV
-  $ erda-cli config delete --orgName xxx --projectName yyy
+  $ erda-cli config delete --org xxx --project yyy  --workspace DEV
+  $ erda-cli config delete --orgN xxx --project yyy
 `,
 	Flags: []command.Flag{
-		command.StringFlag{Name: "orgName", Doc: "[required]which org the project belongs", DefaultValue: ""},
-		command.StringFlag{Name: "projectName", Doc: "[required]which project's feature to delete", DefaultValue: ""},
+		command.StringFlag{Name: "org", Doc: "[required]which org the project belongs", DefaultValue: ""},
+		command.StringFlag{Name: "project", Doc: "[required]which project's feature to delete", DefaultValue: ""},
 		command.StringFlag{Name: "workspace", Doc: "[optional]which workspace's feature to delete", DefaultValue: ""},
 	},
 	Run: RunFeaturesDelete,
 }
 
-func RunFeaturesDelete(ctx *command.Context, orgName, projectName, workspace string) error {
-	var resp apistructs.ExtensionVersionGetResponse
-	var b bytes.Buffer
-
-	if projectName == "" || workspace == "" || orgName == "" {
+func RunFeaturesDelete(ctx *command.Context, org, project, workspace string) error {
+	if project == "" || workspace == "" || org == "" {
 		return fmt.Errorf(
-			utils.FormatErrMsg("config delete", "failed to delete config, one of the flags [orgName, projectName, workspace] not set", true))
+			utils.FormatErrMsg("config delete", "failed to delete config, one of the flags [org, project, workspace] not set", true))
 	}
 
-	uop, err := GetUserOrgProjID(ctx, orgName, projectName)
-	if err != nil {
-		return fmt.Errorf(
-			utils.FormatErrMsg("config delete", "failed to delete config, can not get orgID or userID or projectID: "+err.Error(), true))
-	}
-
-	urlPath := ""
-	if workspace != "" {
-		urlPath = fmt.Sprintf("/api/project-workspace-abilities?projectID=%s&workspace=%s", uop.ProjectId, workspace)
-	} else {
-		urlPath = fmt.Sprintf("/api/project-workspace-abilities?projectID=%s", uop.ProjectId)
-	}
-
-	response, err := ctx.Delete().Path(urlPath).
-		Header(httputil.OrgHeader, uop.OrgId).
-		Do().Body(&b)
-
-	if err != nil {
-		return fmt.Errorf(
-			utils.FormatErrMsg("config delete", "failed to request ("+err.Error()+")", false))
-	}
-
-	if !response.IsOK() {
-		return fmt.Errorf(utils.FormatErrMsg("config delete",
-			fmt.Sprintf("failed to request, status-code: %d, content-type: %s, raw bod: %s",
-				response.StatusCode(), response.ResponseHeader("Content-Type"), b.String()), false))
-	}
-
-	if err = json.Unmarshal(b.Bytes(), &resp); err != nil {
-		return fmt.Errorf(utils.FormatErrMsg("config delete",
-			fmt.Sprintf("failed to unmarshal config delete response ("+err.Error()+")"), false))
-	}
-
-	if !resp.Success {
-		return fmt.Errorf(utils.FormatErrMsg("config delete",
-			fmt.Sprintf("failed to request, error code: %s, error message: %s",
-				resp.Error.Code, resp.Error.Msg), false))
+	if err := common.DelelteProjectWorkspaceConfigs(ctx, org, project, workspace); err != nil {
+		return err
 	}
 
 	ctx.Succ("config delete success\n")
